pyproject.toml
[project]
name = "tmp-pmendbuu2e"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "requests>=2.32.5",
]
